chore: moved analytics segment var into SegmentManager
