+++ OpenEdX/Managers/AnalyticsManager/Services/SegmentAnalyticsService.swift
@@ -7,18 +7,23 @@
 
 import Foundation
 import UIKit
+import Swinject
 
 class SegmentAnalyticsService: AnalyticsService {
     func identify(id: String, username: String?, email: String?) {
-        guard let email = email, let username = username else { return }
+        guard let email = email, 
+              let username = username,
+              let segmentManager = Container.shared.resolve(SegmentManager.self)
+        else { return }
         let traits: [String: String] = [
             "email": email,
             "username": username
         ]
-        (UIApplication.shared.delegate as? AppDelegate)?.analytics?.identify(userId: id, traits: traits)
+        segmentManager.analytics?.identify(userId: id, traits: traits)
     }
     func logEvent(_ event: Event, parameters: [String: Any]?) {
-        (UIApplication.shared.delegate as? AppDelegate)?.analytics?.track(
+        guard let segmentManager = Container.shared.resolve(SegmentManager.self) else { return }
+        segmentManager.analytics?.track(
             name: event.rawValue,
             properties: parameters
         )
